// @types/generated/contentful.d.ts
export interface IBenefitCardFields {
  /** Title */
  title: string

  /** Text */
  text: string

  /** icon */
  icon: 'calendar' | 'airplane' | 'truck'
}

/** A card with an icon, title and text that should describe one benefit of purchasing services from BagBee */

export interface IBenefitCard extends Entry<IBenefitCardFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'benefitCard'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface IFaqSectionFields {
  /** Heading */
  heading: string

  /** Text */
  text: Document
}

/** The FAQ title and text */

export interface IFaqSection extends Entry<IFaqSectionFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'faqSection'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface IFrequentlyAskedQuestionFields {
  /** title */
  title: string

  /** text */
  text: Document
}

export interface IFrequentlyAskedQuestion
  extends Entry<IFrequentlyAskedQuestionFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'frequentlyAskedQuestion'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface IHeroSectionFields {
  /** Title */
  title: string

  /** Subheading */
  subheading: string

  /** CTA Button Text */
  ctaButtonText: string

  /** Hero Background Video */
  heroBackgroundVideo?: Asset | undefined
}

/** The landing page hero section */

export interface IHeroSection extends Entry<IHeroSectionFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'heroSection'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface IImageAndTextFields {
  /** Image */
  image: Asset

  /** Title */
  title: string

  /** Text */
  text: string

  /** CTA Button Text */
  ctaButtonText: string
}

export interface IImageAndText extends Entry<IImageAndTextFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'imageAndText'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface ILandingPageFields {
  /** Title */
  title: string

  /** Hero Section */
  heroSection: IHeroSection

  /** BenefitCards */
  benefitCards: IBenefitCard[]

  /** How It Works Blocks */
  howItWorksBlocks: IImageAndText[]

  /** Video Section */
  videoSection: IVideoSection

  /** FAQ Section */
  faqSection: IFaqSection

  /** FAQ items */
  faq: IFrequentlyAskedQuestion[]

  /** Price Section */
  priceSection: IPriceSection[]

  /** Logo Section Title */
  logoSectionTitle: string

  /** Trustpilot Widget */
  trustpilotWidget: ITrustpilotWidget
}
// ...
export interface IPriceSectionFields {
  /** Title */
  title: string

  /** text */
  text?: string | undefined
}

export interface IPriceSection extends Entry<IPriceSectionFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'priceSection'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}
// ...
export interface ITrustpilotWidgetFields {
  /** Title */
  title: string

  /** Text */
  text: Document

  /** Button text */
  buttonText: string

  /** Button URL */
  buttonUrl: string
}

export interface ITrustpilotWidget extends Entry<ITrustpilotWidgetFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'trustpilotWidget'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}

export interface IVideoSectionFields {
  /** Video */
  video: Asset

  /** Poster */
  poster?: Asset | undefined
}

export interface IVideoSection extends Entry<IVideoSectionFields> {
  sys: {
    id: string
    type: string
    createdAt: string
    updatedAt: string
    locale: string
    contentType: {
      sys: {
        id: 'videoSection'
        linkType: 'ContentType'
        type: 'Link'
      }
    }
  }
}